requirements.txt
arcade==3.3.3
Pillow==11.3.0
pyinstaller==6.16.0